etl/requirements.txt
airflow
apache_airflow_providers_celery
apache-airflow-providers-apache-spark
apache-airflow[google]
pandas
protobuf
pymongo
pyspark
python-dotenv
requests
google-cloud-storage